requirements.txt
numpy
pandas
discord.py
validators
